data_generator/requirements.txt
aiocache==0.11.1
aiofiles==0.4.0
httpx==0.11.1
jinja2==2.11.1
sanpy==0.8.6
starlette==0.13.2
uvicorn==0.11.3
